Margsathi/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
httpx>=0.24.0
python-dotenv>=1.0.0
